update name file and create common css
