+++ src/routes/AppRoutes.js
@@ -3,14 +3,12 @@ import { Routes, Route } from 'react-router-dom';
 
 import Home from '../components/Home/Home.jsx';
 import About from '../components/About/About.jsx';
-import HomeFormEdit from '../components/Backoffice/EditHome/HomeFormEdit'
 
 export default function AppRoutes() {
   return (
     <Routes>
       <Route path="/" element={<Home />} />
       <Route path="/about" element={<About />} />
-      <Route path='/formulario' element={<HomeFormEdit />}/>
     </Routes>
   )
 }
