+++ src/shared/style/baseStyle.css
@@ -1,23 +1,3 @@
-.background_red {
-  background-color: #db5752;
-}
-
-.background_yellow {
-  background-color: #fafa88;
-}
-
-.background_blue {
-  background-color: #9ac9fb;
-}
-
-.background_gray {
-  background-color: #adabab86;
-}
-
-.color_red {
-  color: red;
-}
-
 .button_style {
   background-color: white !important;
   color: #18a0fb !important;
@@ -32,3 +12,10 @@
   background-color: #18a0fb !important;
 }
 
+.background_gray {
+  background-color: #adabab86;
+}
+
+.color_red {
+  color: red;
+}
