+++ src/shared/style/colorOng.css
@@ -1,0 +1,18 @@
+.background_red {
+  background-color: #db5752;
+}
+
+.background_yellow {
+  background-color: #fafa88;
+}
+
+.background_blue {
+  background-color: #9ac9fb;
+}
+
+.color_red_ong {
+  color: #db5752;
+}
+
+
+
